+++ mobile/android/android-components/components/feature/webnotifications/src/main/java/mozilla/components/feature/webnotifications/WebNotificationIntentProcessor.kt
@@ -28,7 +28,7 @@ class WebNotificationIntentProcessor(
         return when (engineNotification) {
             null -> false
             else -> {
-                engine.handleWebNotificationClick(engineNotification)
+                engine.handleWebNotificationClick(engineNotification, intent.action)
                 true
             }
         }